+++ docker/example-server/resources/server/src/de/qtc/rmg/server/legacy/LegacyServer.java
@@ -1,7 +1,6 @@
 package de.qtc.rmg.server.legacy;
 
 import java.net.MalformedURLException;
-import java.rmi.AccessException;
 import java.rmi.AlreadyBoundException;
 import java.rmi.Naming;
 import java.rmi.NotBoundException;
@@ -31,7 +30,8 @@ public class LegacyServer {
         Logger.increaseIndent();
 
         try {
-            Logger.printlnMixedYellow("Creating RMI-Registry on port", String.valueOf(registryPort));
+            Logger.printMixedBlue("Creating", "RMI-Registry", "on port ");
+            Logger.printlnPlainYellow(String.valueOf(registryPort));
             Registry registry = LocateRegistry.createRegistry(registryPort);
             Logger.println("");
 
@@ -52,12 +52,12 @@ public class LegacyServer {
             Logger.printlnMixedBlue("Creating", "PlainServer", "object.");
             remoteObject2 = new PlainServer();
             IPlainServer stub1 = (IPlainServer)UnicastRemoteObject.exportObject(remoteObject2, 0);
-            bindToRegistry(stub1, registry, "plain-server");
+            Utils.bindToRegistry(stub1, registry, "plain-server");
 
             Logger.printlnMixedBlue("Creating another", "PlainServer", "object.");
             remoteObject3 = new PlainServer();
             IPlainServer stub2 = (IPlainServer)UnicastRemoteObject.exportObject(remoteObject3, 0);
-            bindToRegistry(stub2, registry, "plain-server2");
+            Utils.bindToRegistry(stub2, registry, "plain-server2");
 
             try {
                 Logger.printlnMixedBlue("Creating", "ActivatorImp", "object.");
@@ -77,25 +77,12 @@ public class LegacyServer {
             Logger.decreaseIndent();
 
             Logger.println("Server setup finished.");
-            Logger.println("Waiting for incoming connections.");
+            Logger.println("Initializing activation server.");
             Logger.println("");
 
         } catch (RemoteException | MalformedURLException | AlreadyBoundException | NotBoundException e) {
             Logger.eprintln("Unexpected RMI Error:");
             e.printStackTrace();
         }
-    }
-
-    public static void bindToRegistry(Remote object, Registry registry, String boundName) throws AccessException, RemoteException, AlreadyBoundException, NotBoundException
-    {
-        Logger.increaseIndent();
-        Logger.printlnMixedYellow("Binding Object as", boundName);
-        registry.bind(boundName, object);
-
-        Object o = registry.lookup(boundName);
-        String className = o.getClass().getInterfaces()[0].getSimpleName();
-        Logger.printMixedYellow("Boundname", boundName);
-        Logger.printlnPlainMixedBlue(" with interface", className, "is ready.");
-        Logger.decreaseIndent();
     }
 }