+++ CHANGELOG.md
@@ -6,6 +6,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [4.3.0] - May 11, 2022
+
+### Added
+
+* Add support for `ActivatableRef` ([docs](/docs/rmg/actions.md#activatable-bound-names))
+* Add test cases for `ActivatableRef`
+
+### Changed
+
+* Update list of known endpoints ([docs](/docs/rmi/known-endpoints.md))
+* Update outdated documentation
+
+### Docker
+
+* The [example server](/docker/example-server) now provides a full working *Activation System* on port `1098`
+
+
 ## [4.2.2] - Jan 11, 2022
 
 ### Changed